Extract logic to fetch and merge data into a function
